Treat print_footer_scripts like a filter and return the intended value
While debugging the way concatenated scripts works, I discovered that the print_footer_script in bcms (and GigaOM's go-datamodule plugin) were treating the filter as an action.  The function `print_footer_scripts` in `wp-includes/script-loader.php` uses the filter to determine whether or not to fire off `_print_scripts`, which handles concatenation logic.